# Add support for #-style comments