--- a/src/test/php/lang/ast/unittest/TokensTest.class.php
+++ b/src/test/php/lang/ast/unittest/TokensTest.class.php
@@ -40,7 +40,7 @@ class TokensTest extends \unittest\TestCase {
     $this->assertTokens([['string' => $input]], new Tokens(new StringTokenizer($input)));
   }
 
-  #[@test, @expect(class= FormatException::class, withMessage= '/Unclosed string literal/'), @values([
+  #[@test, @expect(['class' => FormatException::class, 'withMessage' => '/Unclosed string literal/']), @values([
   #  '"',
   #  "'",
   #  '"Test',