--- a/src/test/php/lang/ast/unittest/emit/AnnotationsTest.class.php
+++ b/src/test/php/lang/ast/unittest/emit/AnnotationsTest.class.php
@@ -1,6 +1,5 @@
 <?php namespace lang\ast\unittest\emit;
 
-use lang\FormatException;
 use lang\IllegalArgumentException;
 
 /**