Before, these yielded a syntax error.