@thekid 2019-11-17
Comments are not retained inside the emitted code, so this may become a source of confusion if people use XP-style annotations (e.g. `#[@test]`) instead of the Hacklang-inspired `<<test>>`). Previously, there was at least a syntax error, so people would notice.