@thekid 2019-11-23
## Performance

Before:

```bash
$ XP_RT=7.4 xp test src/test/php/
# ...
♥: 577/577 run (0 skipped), 577 succeeded, 0 failed
Memory used: 7818.67 kB (7876.94 kB peak)
Time taken: 0.226 seconds
```

After:

```bash
$ XP_RT=7.4 xp test src/test/php/
# ...
♥: 592/592 run (0 skipped), 592 succeeded, 0 failed
Memory used: 7965.44 kB (8023.70 kB peak)
Time taken: 0.228 seconds
```

## Code

Before:

```bash
$ cloc-1.82.pl src/main/
      51 text files.
      51 unique files.
      10 files ignored.

github.com/AlDanial/cloc v 1.82  T=0.20 s (251.3 files/s, 22685.6 lines/s)
-------------------------------------------------------------------------------
Language                     files          blank        comment           code
-------------------------------------------------------------------------------
PHP                             50            536            684           3293
-------------------------------------------------------------------------------
SUM:                            50            536            684           3293
-------------------------------------------------------------------------------
```

After:

```bash
$ cloc-1.82.pl src/main/
      51 text files.
      51 unique files.
      10 files ignored.

github.com/AlDanial/cloc v 1.82  T=0.23 s (214.1 files/s, 19483.2 lines/s)
-------------------------------------------------------------------------------
Language                     files          blank        comment           code
-------------------------------------------------------------------------------
PHP                             50            540            684           3327
-------------------------------------------------------------------------------
SUM:                            50            540            684           3327
-------------------------------------------------------------------------------
```

...a difference of +34 lines