@thekid 2019-11-23
Handling the key/value pair syntax requires an additional 27 lines of code. Once xp-framework/rfc#335 is implemented, this can be removed.